+++ src/main/java/org/hv/pocket/config/DatabaseConfig.java
@@ -1,16 +1,18 @@
 package org.hv.pocket.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
 
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 /**
  * @author wujianchuan 2019/1/16
  */
-@Component
 @ConfigurationProperties(prefix = "pocket.datasource")
 public class DatabaseConfig {
+
+    private Map<String, DatabaseNodeConfig> nodeConfigMap;
 
     private List<DatabaseNodeConfig> node;
 
@@ -18,7 +20,12 @@ public class DatabaseConfig {
         return node;
     }
 
+    public DatabaseNodeConfig getNode(String nodeName) {
+        return nodeConfigMap.get(nodeName);
+    }
+
     public void setNode(List<DatabaseNodeConfig> node) {
         this.node = node;
+        this.nodeConfigMap = this.node.stream().collect(Collectors.toMap(DatabaseNodeConfig::getNodeName, item -> item));
     }
 }