+++ src/main/java/org/hv/pocket/identify/IncrementStrGenerator.java
@@ -2,19 +2,17 @@ package org.hv.pocket.identify;
 
 import org.hv.pocket.model.AbstractEntity;
 import org.hv.pocket.session.Session;
-import org.springframework.stereotype.Component;
 
 import java.io.Serializable;
 
 /**
  * @author wujianchuan 2019/1/2
  */
-@Component
 public class IncrementStrGenerator extends IncrementLongGenerator {
 
     @Override
     public void setGeneratorId() {
-        this.generationType = GenerationType.STR_INCREMENT;
+        super.generationType = GenerationType.STR_INCREMENT;
     }
 
     @Override